+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change Log
 <a name="v1.3.1"></a>
 ## Version [v1.3.1](https://github.com/luizfp/PrologWebService/compare/v1.3.0...v1.3.1) (release-date) [unreleased]
 
+### Refactor
+* Remove codMarcacaoPorUnidade do IntervaloMarcacao (PL-3199)
 
 <a name="v1.3.0"></a>
 ## Version [v1.3.0](https://github.com/luizfp/PrologWebService/compare/v1.2.1...v1.3.0) (2020-10-13)